refractor  rotation
