+++ parachain/modules/consensus/bnb-pos/verifier/src/primitives.rs
@@ -6,9 +6,6 @@ use anyhow::anyhow;
 use codec::{Decode, Encode};
 use ethabi::ethereum_types::{Bloom, H160, H256, H64, U256};
 use ismp::util::Keccak256;
-use sync_committee_primitives::constants::BlsPublicKey;
-
-use crate::NextValidators;
 
 pub const EPOCH_LENGTH: u64 = 200;
 const EXTRA_VANITY_LENGTH: usize = 32;
@@ -18,14 +15,6 @@ const VALIDATOR_BYTES_LENGTH: usize = 20 + BLS_PUBLIC_KEY_LENGTH;
 const VALIDATOR_NUMBER_SIZE: usize = 1; // // Fixed number of extra prefix bytes reserved for validator number after Luban
 const ADDRESS_LENGTH: usize = 20;
 pub const VALIDATOR_BIT_SET_SIZE: usize = 64;
-
-#[derive(codec::Encode, codec::Decode, Debug)]
-pub struct ConsensusState {
-    pub current_validators: Vec<BlsPublicKey>,
-    pub next_validators: Option<NextValidators>,
-    pub finalized_height: u64,
-    pub finalized_hash: H256,
-}
 
 #[derive(Debug, Encode, Decode)]
 pub struct BnbClientUpdate {
